Refactoring and removing Settings
