+++ Microsoft.Azure.Cosmos/src/Resource/Conflict/CosmosConflict.cs
@@ -1,33 +1,89 @@
-//------------------------------------------------------------
+﻿//------------------------------------------------------------
 // Copyright (c) Microsoft Corporation.  All rights reserved.
 //------------------------------------------------------------
 
 namespace Microsoft.Azure.Cosmos
 {
-    using System.Threading;
-    using System.Threading.Tasks;
+    using System;
+    using System.IO;
+    using Newtonsoft.Json;
+    using Newtonsoft.Json.Converters;
 
     /// <summary>
-    /// This is the conflicting resource resulting from a concurrent async operation in the Azure Cosmos DB service.
+    /// Represents a conflict in the Azure Cosmos DB service.
     /// </summary>
-    /// <remarks>
-    /// On rare occasions, during an async operation (insert, replace and delete), a version conflict may occur on a resource.
-    /// The conflicting resource is persisted as a Conflict resource.  
-    /// Inspecting Conflict resources will allow you to determine which operations and resources resulted in conflicts.
-    /// </remarks>
-    public abstract class CosmosConflict
+    public class CosmosConflict
     {
         /// <summary>
-        /// The Id of the Cosmos conflict
+        /// Gets or sets the Id of the resource in the Azure Cosmos DB service.
         /// </summary>
-        public abstract string Id { get; }
+        /// <value>The Id associated with the resource.</value>
+        /// <remarks>
+        /// <para>
+        /// Every resource within an Azure Cosmos DB database account needs to have a unique identifier. 
+        /// </para>
+        /// <para>
+        /// The following characters are restricted and cannot be used in the Id property:
+        ///  '/', '\\', '?', '#'
+        /// </para>
+        /// </remarks>
+        [JsonProperty(PropertyName = Documents.Constants.Properties.Id)]
+        public virtual string Id { get; set; }
 
         /// <summary>
-        /// Deletes the current conflict instance
+        /// Gets or sets the operation that resulted in the conflict in the Azure Cosmos DB service.
         /// </summary>
-        /// <param name="partitionKey">The partition key for the item.</param>
-        /// <param name="cancellationToken">(Optional) <see cref="CancellationToken"/> representing request cancellation.</param>
+        [JsonConverter(typeof(StringEnumConverter))]
+        [JsonProperty(PropertyName = Documents.Constants.Properties.OperationType)]
+        public virtual OperationKind OperationKind { get; set; }
+
+        /// <summary>
+        /// Gets or sets the type of the conflicting resource in the Azure Cosmos DB service.
+        /// </summary>
+        [JsonConverter(typeof(ConflictResourceTypeJsonConverter))]
+        [JsonProperty(PropertyName = Documents.Constants.Properties.ResourceType)]
+        public virtual Type ResourceType { get; set; }
+
+        /// <summary>
+        /// Gets the resource ID for the conflict in the Azure Cosmos DB service.
+        /// </summary>
+        [JsonProperty(PropertyName = Documents.Constants.Properties.SourceResourceId)]
+        public virtual string SourceResourceId { get; set; }
+
+        /// <summary>
+        /// Gets the content of the Conflict resource in the Azure Cosmos DB service.
+        /// </summary>
+        /// <typeparam name="T">The type to use to deserialize the content.</typeparam>
+        /// <param name="cosmosJsonSerializer">(Optional) <see cref="CosmosJsonSerializer"/> to use while parsing the content.</param>
         /// <returns></returns>
-        public abstract Task<CosmosConflictResponse> DeleteAsync(object partitionKey, CancellationToken cancellationToken = default(CancellationToken));
+        public virtual T GetResource<T>(CosmosJsonSerializer cosmosJsonSerializer = null)
+        {
+            if (!string.IsNullOrEmpty(this.Content))
+            {
+                if (cosmosJsonSerializer == null)
+                {
+                    cosmosJsonSerializer = new CosmosDefaultJsonSerializer();
+                }
+
+                using (MemoryStream stream = new MemoryStream())
+                {
+                    using (StreamWriter writer = new StreamWriter(stream))
+                    {
+                        writer.Write(this.Content);
+                        writer.Flush();
+                        stream.Position = 0;
+                        return cosmosJsonSerializer.FromStream<T>(stream);
+                    }
+                }
+            }
+
+            return default(T);
+        }
+
+        [JsonProperty(PropertyName = Documents.Constants.Properties.Content)]
+        internal string Content { get; set; }
+
+        [JsonProperty(PropertyName = Documents.Constants.Properties.ConflictLSN)]
+        internal long ConflictLSN { get; set; }
     }
 }
