+++ Microsoft.Azure.Cosmos/src/Resource/Conflict/CosmosConflicts.cs
@@ -10,18 +10,12 @@ namespace Microsoft.Azure.Cosmos
     public abstract class CosmosConflicts
     {
         /// <summary>
-        /// Returns a reference to a conflict object. 
-        /// </summary>
-        /// <param name="id">The cosmos conflict id.</param>
-        public abstract CosmosConflict this[string id] { get; }
-
-        /// <summary>
-        /// Obtains an iterator to go through the <see cref="CosmosConflictSettings"/> on an Azure Cosmos container.
+        /// Obtains an iterator to go through the <see cref="CosmosConflict"/> on an Azure Cosmos container.
         /// </summary>
         /// <param name="maxItemCount">(Optional) The max item count to return as part of the query</param>
         /// <param name="continuationToken">(Optional) The continuation token in the Azure Cosmos DB service.</param>
         /// <returns></returns>
-        public abstract CosmosResultSetIterator<CosmosConflictSettings> GetConflictsIterator(
+        public abstract CosmosResultSetIterator<CosmosConflict> GetConflictsIterator(
             int? maxItemCount = null,
             string continuationToken = null);
