+++ Microsoft.Azure.Cosmos/src/Resource/Settings/CosmosContainerSettings.cs
@@ -176,9 +176,7 @@ namespace Microsoft.Azure.Cosmos
 
             set
             {
-                if (value == null) throw new ArgumentNullException($"{nameof(value)}");
-
-                this.uniqueKeyPolicyInternal = value;
+                this.uniqueKeyPolicyInternal = value ?? throw new ArgumentNullException($"{nameof(value)}");
             }
         }
